requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.1
openai==1.12.0
pydantic==2.6.1
pytz==2024.1
google-search-results==2.4.2